got the model running!

+++ main.py
@@ -48,7 +48,7 @@ print(TList)
 new_model = OverallModel(FnList,VnList,TList)
 
 simDur =  30 * 24 * 60 * 60 #number of seconds to run sim for
-dts = simDur/new_model.time_step
+dts = int(simDur/new_model.time_step)
 print("Beginning Simulation")
 for i in range(0,dts):
     new_model.step()
